@@ -13,38 +13,43 @@
     "lint:fix": "./node_modules/eslint/bin/eslint.js src --fix",
     "flow": "./node_modules/flow-bin/cli.js",
     "e2e:build": "detox build --configuration ios.sim.debug",
-    "e2e:test": "detox test --configuration ios.sim.debug"
+    "e2e:test": "detox test --configuration ios.sim.debug",
+    "postinstall": "npx jetify"
   },
   "dependencies": {
+    "@react-native-firebase/analytics": "^6.0.3",
+    "@react-native-firebase/app": "^6.0.3",
     "i18n-js": "^3.1.0",
     "lodash": "^4.17.11",
     "moment": "^2.22.2",
     "moment-timezone": "^0.5.23",
-    "react": "16.6.1",
-    "react-native": "0.57.5",
-    "react-native-calendar-events": "^1.6.3",
-    "react-native-camera": "gaykov/react-native-camera",
+    "react": "16.9.0",
+    "react-native": "0.61.2",
+    "react-native-calendar-events": "^1.7.3",
+    "react-native-camera": "^3.8.0",
     "react-native-color": "^0.0.10",
-    "react-native-contacts": "^2.2.4",
-    "react-native-easy-toast": "gaykov/react-native-easy-toast",
-    "react-native-firebase": "^5.1.1",
-    "react-native-gesture-handler": "^1.0.9",
-    "react-native-iap": "2.4.0-beta6",
-    "react-native-keyboard-aware-scroll-view": "^0.7.4",
-    "react-native-languages": "^3.0.1",
-    "react-native-linear-gradient": "^2.5.2",
-    "react-native-open-maps": "^0.3.2",
-    "react-native-paper": "^2.3.0",
-    "react-native-qrcode-scanner": "^1.1.0",
-    "react-native-qrcode-svg": "^5.1.1",
-    "react-native-rate": "^1.1.6",
-    "react-native-share": "^1.1.3",
-    "react-native-sound": "^0.10.9",
-    "react-native-svg": "^8.0.9",
-    "react-native-ui-lib": "^3.5.0",
-    "react-native-vector-icons": "^6.1.0",
-    "react-navigation": "^3.0.0",
-    "react-navigation-material-bottom-tabs": "^1.0.0",
+    "react-native-contacts": "^5.0.4",
+    "react-native-easy-toast": "^1.2.0",
+    "react-native-gesture-handler": "^1.4.1",
+    "react-native-iap": "^4.0.8",
+    "react-native-keyboard-aware-scroll-view": "^0.9.1",
+    "react-native-linear-gradient": "^2.5.6",
+    "react-native-localize": "^1.3.1",
+    "react-native-open-maps": "^0.3.4",
+    "react-native-paper": "^3.0.0",
+    "react-native-qrcode-svg": "^5.2.0",
+    "react-native-rate": "^1.1.10",
+    "react-native-reanimated": "^1.3.0",
+    "react-native-screens": "^1.0.0-alpha.23",
+    "react-native-share": "^2.0.0",
+    "react-native-sound": "0.11.0",
+    "react-native-svg": "^9.13.3",
+    "react-native-ui-lib": "^3.39.0",
+    "react-native-vector-icons": "^6.6.0",
+    "react-navigation": "^4.0.10",
+    "react-navigation-material-bottom-tabs": "^2.1.3",
+    "react-navigation-stack": "^1.10.3",
+    "react-navigation-tabs": "^2.5.6",
     "react-redux": "^5.1.1",
     "recompose": "^0.30.0",
     "redux": "^4.0.1",
@@ -55,26 +60,26 @@
     "vcard-parser": "^1.0.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.6.2",
     "@babel/preset-flow": "^7.0.0",
-    "babel-core": "^7.0.0-bridge",
+    "@babel/runtime": "^7.6.2",
+    "@react-native-community/eslint-config": "^0.0.5",
     "babel-eslint": "^10.0.1",
-    "babel-jest": "23.6.0",
-    "detox": "^12.9.0",
-    "enzyme": "^3.7.0",
-    "enzyme-adapter-react-16": "^1.7.0",
-    "eslint": "^5.9.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "^6.1.2",
-    "eslint-plugin-react": "^7.11.1",
-    "flow-bin": "0.78.0",
-    "jest": "23.6.0",
+    "babel-jest": "^24.9.0",
+    "enzyme": "^3.9.0",
+    "enzyme-adapter-react-16": "^1.11.2",
+    "enzyme-to-json": "^3.3.5",
+    "eslint": "^6.5.1",
+    "eslint-config-prettier": "^6.4.0",
+    "flow-bin": "^0.105.0",
+    "jest": "^24.9.0",
+    "jetifier": "^1.6.4",
     "jsdom": "13.0.0",
     "jsdom-global": "3.0.2",
-    "metro-react-native-babel-preset": "0.49.2",
+    "metro-react-native-babel-preset": "^0.56.0",
     "react-addons-test-utils": "^15.6.2",
     "react-dom": "^16.6.3",
-    "react-test-renderer": "16.6.1",
+    "react-test-renderer": "16.9.0",
     "sinon": "^7.1.1"
   },
   "jest": {
