@@ -1,6 +1,3 @@
-module.exports = function (api) {
-  api.cache(true);
-  return {
-    presets: ['module:metro-react-native-babel-preset', '@babel/flow'],
-  };
+module.exports = {
+  presets: ['module:metro-react-native-babel-preset'],
 };
