@@ -1,8 +1,6 @@
 // @flow
 import React from 'react';
-import {
-  Modal, Picker, DatePickerIOS, Platform,
-} from 'react-native';
+import { Modal, Picker, DatePickerIOS, Platform } from 'react-native';
 import { View, Text } from 'react-native-ui-lib';
 import moment from 'moment';
 
@@ -11,6 +9,7 @@ import { Input, Button } from '../index';
 /**
  * Props type for PickerComponent
  */
+
 export type PickerProps = {
   /**
    * Picker type
@@ -35,8 +34,8 @@ export type PickerProps = {
   /**
    * Function to handle inputValue change
    */
-  handleSetValue: (Date) => void,
-  setInputValue: (string) => void,
+  handleSetValue: Date => void,
+  setInputValue: string => void,
   inputValue: Date,
 
   minimumDate: any,
@@ -44,7 +43,7 @@ export type PickerProps = {
   /**
    * Callback it value has set
    */
-  onSetValue: (string) => void,
+  onSetValue: string => void,
 
   items?: Array<{
     id: number,
@@ -62,13 +61,16 @@ const PickerComponent = (props: PickerProps) => (
         <Input
           key={props.key}
           placeholder={props.placeholder}
-          onFocus={Platform.OS === 'android' && props.type === 'datetime'
-            ? props.openAndroidDatepicker
-            : props.toggleModal
+          onFocus={
+            Platform.OS === 'android' && props.type === 'datetime'
+              ? props.openAndroidDatepicker
+              : props.toggleModal
           }
-          value={props.type === 'datetime'
-            ? props.inputValue && moment(props.inputValue).format('DD-MM-YYYY hh:mm')
-            : props.inputValue
+          value={
+            props.type === 'datetime'
+              ? props.inputValue &&
+                moment(props.inputValue).format('DD-MM-YYYY hh:mm')
+              : props.inputValue
           }
         />
         <Modal
@@ -77,43 +79,57 @@ const PickerComponent = (props: PickerProps) => (
           visible={props.isModalOpened}
         >
           <View paddingH-15 flex-1 centerV>
-            <Text h2 center marginB-20>{props.title}</Text>
-            {props.type === 'datetime'
-              ? (
-                <DatePickerIOS
-                  date={props.inputValue ? props.inputValue : new Date()}
-                  onDateChange={props.handleSetValue}
-                  minimumDate={props.minimumDate}
-                />
-              ) : (
-                <Picker
-                  selectedValue={props.inputValue}
-                  onValueChange={props.handleSetValue}
-                >
-                  {props.items && props.items.map(item => <Picker.Item key={item.id} {...item} />)}
-                </Picker>
-              )
-            }
-            {Platform.OS === 'ios' && <Button onPress={props.toggleModal} style={{ marginTop: 20 }}>Save</Button>}
+            <Text h2 center marginB-20>
+              {props.title}
+            </Text>
+            {props.type === 'datetime' ? (
+              <DatePickerIOS
+                date={props.inputValue ? props.inputValue : new Date()}
+                onDateChange={props.handleSetValue}
+                minimumDate={props.minimumDate}
+              />
+            ) : (
+              <Picker
+                selectedValue={props.inputValue}
+                onValueChange={props.handleSetValue}
+              >
+                {props.items &&
+                  props.items.map(item => (
+                    <Picker.Item key={item.id} {...item} />
+                  ))}
+              </Picker>
+            )}
+            {Platform.OS === 'ios' && (
+              <Button onPress={props.toggleModal} style={{ marginTop: 20 }}>
+                Save
+              </Button>
+            )}
           </View>
         </Modal>
       </View>
     ) : (
       <View row>
-        {props.items && props.items.map((item, index) => (
-          // $$FlowFixMe
-          <View flex marginR-20={props.items.length - 1 !== index} key={item.id}>
-            <Button
-              radius={5}
-              onPress={() => props.handleSetValue(item.value)}
-              variant={props.inputValue === item.value ? 'darkBlue' : 'lightGray'}
+        {props.items &&
+          props.items.map((item, index) => (
+            // $$FlowFixMe
+            <View
+              flex
+              marginR-20={props.items.length - 1 !== index}
+              key={item.id}
             >
-              <Text white={props.inputValue === item.value} center default>
-                {item.label.length < 5 ? item.label : item.value}
-              </Text>
-            </Button>
-          </View>
-        ))}
+              <Button
+                radius={5}
+                onPress={() => props.handleSetValue(item.value)}
+                variant={
+                  props.inputValue === item.value ? 'darkBlue' : 'lightGray'
+                }
+              >
+                <Text white={props.inputValue === item.value} center default>
+                  {item.label.length < 5 ? item.label : item.value}
+                </Text>
+              </Button>
+            </View>
+          ))}
       </View>
     )}
   </View>
