@@ -2,8 +2,4 @@ import Button from './Button';
 import Input from './Input';
 import Picker from './Picker/PickerContaner';
 
-export {
-  Button,
-  Input,
-  Picker,
-};
+export { Button, Input, Picker };
