@@ -3,10 +3,5 @@ import React from 'react';
 import Navigator from './navigation/Navigator';
 
 export default function AppView() {
-  return (
-    <Navigator
-      onNavigationStateChange={() => {}}
-      uriPrefix="/app"
-    />
-  );
+  return <Navigator onNavigationStateChange={() => {}} uriPrefix="/app" />;
 }
