@@ -17,18 +17,26 @@ export const generateQRValueFromState = (state: State): string => {
 ?subject=${fieldValues[fieldTypesList.EMAIL_SUBJECT] || ''}
 &body=${fieldValues[fieldTypesList.EMAIL_BODY] || ''}`;
     case codeTypesList.SMS:
-      return `SMS:${fieldValues[fieldTypesList.SMS_TO] || ''}:${fieldValues[fieldTypesList.SMS_MESSAGE] || ''}`;
+      return `SMS:${fieldValues[fieldTypesList.SMS_TO] || ''}:${fieldValues[
+        fieldTypesList.SMS_MESSAGE
+      ] || ''}`;
     case codeTypesList.WIFI:
       return `WIFI:S:${fieldValues[fieldTypesList.WIFI_SSID] || ''};
 T:${fieldValues[fieldTypesList.WIFI_ENCRYPTION] || ''};
 P:${fieldValues[fieldTypesList.WIFI_PASSWORD] || ''};;`;
     case codeTypesList.GEO:
-      return `geo:${fieldValues[fieldTypesList.GEO_LONG] || ''},${fieldValues[fieldTypesList.GEO_LAT] || ''}`;
+      return `geo:${fieldValues[fieldTypesList.GEO_LONG] || ''},${fieldValues[
+        fieldTypesList.GEO_LAT
+      ] || ''}`;
     case codeTypesList.CONTACT:
       return `BEGIN:VCARD
 VERSION:4.0
-N:${fieldValues[fieldTypesList.CONTACT_SURNAME] || ''};${fieldValues[fieldTypesList.CONTACT_NAME] || ''};;;
-FN:${fieldValues[fieldTypesList.CONTACT_NAME] || ''} ${fieldValues[fieldTypesList.CONTACT_SURNAME] || ''}
+N:${fieldValues[fieldTypesList.CONTACT_SURNAME] || ''};${fieldValues[
+        fieldTypesList.CONTACT_NAME
+      ] || ''};;;
+FN:${fieldValues[fieldTypesList.CONTACT_NAME] || ''} ${fieldValues[
+        fieldTypesList.CONTACT_SURNAME
+      ] || ''}
 TEL:${fieldValues[fieldTypesList.CONTACT_PHONE] || ''}
 EMAIL:${fieldValues[fieldTypesList.CONTACT_EMAIL] || ''}
 END:VCARD`;
@@ -37,14 +45,20 @@ END:VCARD`;
 SUMMARY:${fieldValues[fieldTypesList.EVENT_TITLE] || ''}
 LOCATION:${fieldValues[fieldTypesList.EVENT_LOCATION] || ''}
 DESCRIPTION:${fieldValues[fieldTypesList.EVENT_DESCRIPTION] || ''}
-DTSTART:${fieldValues[fieldTypesList.EVENT_START]
-    ? moment(fieldValues[fieldTypesList.EVENT_START]).format('YYYYMMDDThhmmss')
-    : ''
-}
-DTEND:${fieldValues[fieldTypesList.EVENT_START]
-    ? moment(fieldValues[fieldTypesList.EVENT_END]).format('YYYYMMDDThhmmss')
-    : ''
-}
+DTSTART:${
+        fieldValues[fieldTypesList.EVENT_START]
+          ? moment(fieldValues[fieldTypesList.EVENT_START]).format(
+              'YYYYMMDDThhmmss',
+            )
+          : ''
+      }
+DTEND:${
+        fieldValues[fieldTypesList.EVENT_START]
+          ? moment(fieldValues[fieldTypesList.EVENT_END]).format(
+              'YYYYMMDDThhmmss',
+            )
+          : ''
+      }
 END:VEVENT`;
     default:
       return fieldValues[fieldTypesList.TEXT] || '';
@@ -54,12 +68,16 @@ END:VEVENT`;
 type GeneratedCodeState = {};
 
 type Action = {
-  type: string, payload: any,
+  type: string,
+  payload: any,
 };
 
 export const initialState: GeneratedCodeState = {};
 
-export default function GeneratedReducer(state: GeneratedCodeState = initialState, action: Action): GeneratedCodeState {
+export default function GeneratedReducer(
+  state: GeneratedCodeState = initialState,
+  action: Action,
+): GeneratedCodeState {
   switch (action.type) {
     default:
       return state;
