@@ -7,11 +7,7 @@ import {
   TouchableOpacity,
   StyleSheet,
 } from 'react-native';
-import {
-  View,
-  Text,
-  Image,
-} from 'react-native-ui-lib';
+import { View, Text, Image } from 'react-native-ui-lib';
 import LinearGradient from 'react-native-linear-gradient';
 import moment from 'moment';
 
@@ -30,7 +26,7 @@ type Props = {
     data: string,
     date: Date,
   }>,
-  removeItemFromHistory: (number) => void,
+  removeItemFromHistory: number => void,
   handleClearHistory: () => void,
   goPricingPage: () => void,
   isPro: boolean,
@@ -40,30 +36,36 @@ export default function HistoryView(props: Props) {
   const itemsToMap = props.isPro ? props.items : props.items.slice(0, 10);
   return (
     <SafeAreaView style={commonStyles.safeArea}>
-      <StatusBar
-        translucent={false}
-        backgroundColor={colors.lightBlue}
-      />
+      <StatusBar translucent={false} backgroundColor={colors.lightBlue} />
       <View row centerH centerV marginB-25 marginT-10>
-        <Text h1 darkBlue>{i18n.t('screens.history.title', { defaultValue: i18n.t('tabs.history') })}</Text>
+        <Text h1 darkBlue>
+          {i18n.t('screens.history.title', {
+            defaultValue: i18n.t('tabs.history'),
+          })}
+        </Text>
         {props.items.length > 10 && (
-          <TouchableOpacity style={styles.clearButton} onPress={props.handleClearHistory}>
+          <TouchableOpacity
+            style={styles.clearButton}
+            onPress={props.handleClearHistory}
+          >
             <Text red>{i18n.t('screens.history.clear')}</Text>
           </TouchableOpacity>
         )}
       </View>
-      <ScrollView
-        style={styles.historyItemsWrapper}
-      >
+      <ScrollView style={styles.historyItemsWrapper}>
         {props.items.length === 0 && (
           <Text marginT-20 default gray center>
             {i18n.t('screens.history.noCodesMessage')}
           </Text>
         )}
-        {itemsToMap.map((qrcode) => {
+        {itemsToMap.map(qrcode => {
           const parsedQRCode = parseScannedString(qrcode.data);
-          const foundCodeType = codeTypes.find(codeType => codeType.label === parsedQRCode.type);
-          const whiteIcon = foundCodeType ? `${foundCodeType.icon}-white` : 'text-white';
+          const foundCodeType = codeTypes.find(
+            codeType => codeType.label === parsedQRCode.type,
+          );
+          const whiteIcon = foundCodeType
+            ? `${foundCodeType.icon}-white`
+            : 'text-white';
           return (
             <TouchableOpacity
               onPress={() => props.navigation.navigate('ScannedCode', qrcode)}
@@ -72,31 +74,28 @@ export default function HistoryView(props: Props) {
             >
               <LinearGradient
                 style={styles.imageContainer}
-                colors={[colors.primaryGradientStart, colors.primaryGradientEnd]}
+                colors={[
+                  colors.primaryGradientStart,
+                  colors.primaryGradientEnd,
+                ]}
                 start={{ x: 0, y: 1 }}
                 end={{ x: 1, y: 0 }}
               >
-                <Image
-                  assetGroup="types"
-                  assetName={whiteIcon}
-                />
+                <Image assetGroup="types" assetName={whiteIcon} />
               </LinearGradient>
               <View paddingH-15 flex-1>
                 <View row centerV spread>
                   <Text numberOfLines={1} h3 marginB-5 marginR-15>
                     {parsedQRCode.fields[0].value}
-                    {(parsedQRCode.type === codeTypesList.CONTACT
-                      || parsedQRCode.type === codeTypesList.GEO)
-                      && ` ${parsedQRCode.fields[1].value}`}
+                    {(parsedQRCode.type === codeTypesList.CONTACT ||
+                      parsedQRCode.type === codeTypesList.GEO) &&
+                      ` ${parsedQRCode.fields[1].value}`}
                   </Text>
                   <TouchableOpacity
                     onPress={() => props.removeItemFromHistory(qrcode.id)}
                     style={styles.deleteIconContainer}
                   >
-                    <Image
-                      assetGroup="icons"
-                      assetName="delete"
-                    />
+                    <Image assetGroup="icons" assetName="delete" />
                   </TouchableOpacity>
                 </View>
                 <Text gray>{moment(qrcode.date).format('DD.MM.YYYY')}</Text>
@@ -105,7 +104,10 @@ export default function HistoryView(props: Props) {
           );
         })}
         {!props.isPro && props.items.length > 10 && (
-          <TouchableOpacity style={styles.needMoreButton} onPress={props.goPricingPage}>
+          <TouchableOpacity
+            style={styles.needMoreButton}
+            onPress={props.goPricingPage}
+          >
             <Text gray>{i18n.t('screens.history.needMore')}&nbsp;</Text>
             <View style={styles.proLabel} paddingH-3 paddingV-1 br20 marginB-4>
               <Text white>{i18n.t('other.pro')}</Text>
