@@ -36,16 +36,19 @@ export const codeTypesList = {
   WIFI: i18n.t('codeTypes.wifi'),
 };
 
-type NewCodeState = {
-};
+type NewCodeState = {};
 
 type Action = {
-  type: string, payload: any,
+  type: string,
+  payload: any,
 };
 
 const initialState: NewCodeState = {};
 
-export default function NewCodeReducer(state: NewCodeState = initialState, action: Action): NewCodeState {
+export default function NewCodeReducer(
+  state: NewCodeState = initialState,
+  action: Action,
+): NewCodeState {
   switch (action.type) {
     default:
       return state;
