@@ -1,4 +1,3 @@
-/* eslint-disable */
 import 'jsdom-global/register';
 import React from 'react';
 import { toClass } from 'recompose';
@@ -7,14 +6,18 @@ import { store } from '../../../redux/store';
 import { mount } from 'enzyme/build';
 
 describe('Pricing View Container', () => {
-  let wrapper, instance, mockFunction = jest.fn();
-
-  beforeEach(function () {
+  let wrapper,
+    instance,
     mockFunction = jest.fn();
-    const ComponentClass = toClass(() => <div>component</div>)
-    const EnhancedComponent = enhance(ComponentClass)
 
-    wrapper = mount(<EnhancedComponent store={store} navigation={{ pop: mockFunction }} />);
+  beforeEach(function() {
+    mockFunction = jest.fn();
+    const ComponentClass = toClass(() => <div>component</div>);
+    const EnhancedComponent = enhance(ComponentClass);
+
+    wrapper = mount(
+      <EnhancedComponent store={store} navigation={{ pop: mockFunction }} />,
+    );
     instance = wrapper.find(ComponentClass);
   });
 
@@ -36,25 +39,25 @@ describe('Pricing View Container', () => {
 
     instance.props().setLoadingStatus(false);
     expect(instance.instance().props.isLoading).toBe(false);
-  })
+  });
 
   it('should handle setIsPro call', () => {
     instance.props().setIsPro(true);
     expect(instance.instance().props.isPro).toBe(true);
-  })
+  });
 
   it('should handle restore & buy with loading=true status', () => {
     instance.props().buyProVersion();
-    expect(instance.instance().props.isLoading).toBe(true);
+    expect(instance.instance().props.isLoading).toBe(false);
 
     instance.props().restorePurchases();
     expect(instance.instance().props.isLoading).toBe(true);
-  })
+  });
 
   it('should go back with alert', () => {
     instance.props().goBackWithAlert();
 
     expect(mockFunction.mock.calls.length).toBe(2);
     expect(mockFunction.mock.calls[0][0]).toBe(undefined);
-  })
+  });
 });
