@@ -10,10 +10,7 @@ import {
   Linking,
   Platform,
 } from 'react-native';
-import {
-  View,
-  Text,
-} from 'react-native-ui-lib';
+import { View, Text } from 'react-native-ui-lib';
 import tinycolor from 'tinycolor2';
 import { SlidersColorPicker } from 'react-native-color';
 import Rate from 'react-native-rate';
@@ -21,10 +18,9 @@ import Rate from 'react-native-rate';
 import { commonStyles, colors } from '../../styles';
 import i18n from '../../translations';
 
-
 type Props = {
   navigation: {
-    navigate: (string) => void,
+    navigate: string => void,
   },
   settings: {
     batch: boolean,
@@ -40,50 +36,57 @@ type Props = {
   isForegroundModalVisible: boolean,
   toggleBackgroundColorModal: () => void,
   toggleForegroundColorModal: () => void,
-  handleBackgroundColorPick: (string) => void,
-  handleForegroundColorPick: (string) => void,
+  handleBackgroundColorPick: string => void,
+  handleForegroundColorPick: string => void,
   setSettingValue: (string, any) => void,
   isPro: boolean,
 };
 
 export const generateCodePreviewColors = (color: string) => ({
   backgroundColor: color,
-  borderColor: tinycolor(color).isLight()
-    ? colors.lightGray
-    : colors.black,
+  borderColor: tinycolor(color).isLight() ? colors.lightGray : colors.black,
 });
 
 export const SettingSwitch = (props: {
   value: boolean,
   onChange: (string, boolean) => void,
   name: string,
-  disabled?: boolean}) => (
-    <Switch
-      value={props.value}
-      onValueChange={value => props.onChange(props.name, value)}
-      disabled={props.disabled}
-      testID="switch"
-      trackColor={{
-        true: Platform.select({
-          android: colors.primaryGradientEnd,
-          ios: colors.lightGray,
-        }),
-        false: colors.lightGray,
-      }}
-      thumbColor={colors.primaryGradientEnd}
-    />
+  disabled?: boolean,
+}) => (
+  <Switch
+    value={props.value}
+    onValueChange={value => props.onChange(props.name, value)}
+    disabled={props.disabled}
+    testID="switch"
+    trackColor={{
+      true: Platform.select({
+        android: colors.primaryGradientEnd,
+        ios: colors.lightGray,
+      }),
+      false: colors.lightGray,
+    }}
+    thumbColor={colors.primaryGradientEnd}
+  />
 );
 
 export default function SettingsView(props: Props) {
-  const recommendedColors = ['#247ba0', '#70c1b3', '#b2dbbf', '#f3ffbd', '#ff1654'];
+  const recommendedColors = [
+    '#247ba0',
+    '#70c1b3',
+    '#b2dbbf',
+    '#f3ffbd',
+    '#ff1654',
+  ];
   return (
     <SafeAreaView style={[commonStyles.safeArea, styles.viewContainer]}>
-      <StatusBar
-        translucent={false}
-        backgroundColor={colors.lightBlue}
-      />
+      <StatusBar translucent={false} backgroundColor={colors.lightBlue} />
       <View centerH marginB-25 marginT-10>
-        <Text h1 darkBlue> {i18n.t('screens.settings.title', { defaultValue: i18n.t('tabs.settings') })}</Text>
+        <Text h1 darkBlue>
+          {' '}
+          {i18n.t('screens.settings.title', {
+            defaultValue: i18n.t('tabs.settings'),
+          })}
+        </Text>
       </View>
       <ScrollView
         style={styles.scrollView}
@@ -94,108 +97,187 @@ export default function SettingsView(props: Props) {
           <View marginB-25>
             <View row spread>
               <View flex>
-                <Text h2 marginB-5 style={styles.textNormal}>{i18n.t('screens.settings.offer')}</Text>
-                <Text gray defaultLight>{i18n.t('screens.settings.offerDescription')}</Text>
+                <Text h2 marginB-5 style={styles.textNormal}>
+                  {i18n.t('screens.settings.offer')}
+                </Text>
+                <Text gray defaultLight>
+                  {i18n.t('screens.settings.offerDescription')}
+                </Text>
               </View>
-              <TouchableOpacity
-                onPress={props.goPricingPage}
-              >
-                <Text h3 primary>{i18n.t('screens.settings.offerButton')}</Text>
+              <TouchableOpacity onPress={props.goPricingPage}>
+                <Text h3 primary>
+                  {i18n.t('screens.settings.offerButton')}
+                </Text>
               </TouchableOpacity>
             </View>
           </View>
         )}
         <View marginB-25>
-          <Text h2 marginB-10>{i18n.t('screens.settings.customization')}</Text>
-          <View br10 paddingH-15 paddingT-25 paddingB-0 style={styles.settingBlock}>
+          <Text h2 marginB-10>
+            {i18n.t('screens.settings.customization')}
+          </Text>
+          <View
+            br10
+            paddingH-15
+            paddingT-25
+            paddingB-0
+            style={styles.settingBlock}
+          >
             <View marginB-25>
               <View row spread>
                 <View row centerV>
-                  <Text h2 marginB-5 style={styles.textNormal}>{i18n.t('screens.settings.background')}</Text>
+                  <Text h2 marginB-5 style={styles.textNormal}>
+                    {i18n.t('screens.settings.background')}
+                  </Text>
                   {!props.isPro && (
-                    <View style={styles.proLabel} paddingH-3 paddingV-1 br20 marginB-4 marginL-5>
-                      <Text white default>{i18n.t('other.pro')}</Text>
+                    <View
+                      style={styles.proLabel}
+                      paddingH-3
+                      paddingV-1
+                      br20
+                      marginB-4
+                      marginL-5
+                    >
+                      <Text white default>
+                        {i18n.t('other.pro')}
+                      </Text>
                     </View>
                   )}
                 </View>
                 <TouchableOpacity
-                  onPress={!props.isPro ? props.goPricingPage : props.toggleBackgroundColorModal}
+                  onPress={
+                    !props.isPro
+                      ? props.goPricingPage
+                      : props.toggleBackgroundColorModal
+                  }
                   style={[
                     styles.colorPreview,
                     generateCodePreviewColors(props.settings.backgroundColor),
                   ]}
                 />
               </View>
-              <Text gray defaultLight>{i18n.t('screens.settings.backgroundDescription')}</Text>
+              <Text gray defaultLight>
+                {i18n.t('screens.settings.backgroundDescription')}
+              </Text>
             </View>
             <View marginB-25>
               <View row spread>
                 <View row centerV>
-                  <Text h2 marginB-5 style={styles.textNormal}>{i18n.t('screens.settings.foreground')}</Text>
+                  <Text h2 marginB-5 style={styles.textNormal}>
+                    {i18n.t('screens.settings.foreground')}
+                  </Text>
                   {!props.isPro && (
-                    <View style={styles.proLabel} paddingH-3 paddingV-1 br20 marginB-4 marginL-5>
-                      <Text white default>{i18n.t('other.pro')}</Text>
+                    <View
+                      style={styles.proLabel}
+                      paddingH-3
+                      paddingV-1
+                      br20
+                      marginB-4
+                      marginL-5
+                    >
+                      <Text white default>
+                        {i18n.t('other.pro')}
+                      </Text>
                     </View>
                   )}
                 </View>
                 <TouchableOpacity
-                  onPress={!props.isPro ? props.goPricingPage : props.toggleForegroundColorModal}
+                  onPress={
+                    !props.isPro
+                      ? props.goPricingPage
+                      : props.toggleForegroundColorModal
+                  }
                   style={[
                     styles.colorPreview,
                     generateCodePreviewColors(props.settings.foregroundColor),
                   ]}
                 />
               </View>
-              <Text gray defaultLight>{i18n.t('screens.settings.foregroundDescription')}</Text>
+              <Text gray defaultLight>
+                {i18n.t('screens.settings.foregroundDescription')}
+              </Text>
             </View>
           </View>
         </View>
         <View marginB-25>
-          <Text h2 marginB-10>{i18n.t('screens.settings.general')}</Text>
-          <View br10 paddingH-15 paddingT-25 paddingB-0 style={styles.settingBlock}>
+          <Text h2 marginB-10>
+            {i18n.t('screens.settings.general')}
+          </Text>
+          <View
+            br10
+            paddingH-15
+            paddingT-25
+            paddingB-0
+            style={styles.settingBlock}
+          >
             <View marginB-25>
               <View row spread centerV>
                 <View row centerV>
-                  <Text h2 marginB-5 style={styles.textNormal}>{i18n.t('screens.settings.batch')}</Text>
+                  <Text h2 marginB-5 style={styles.textNormal}>
+                    {i18n.t('screens.settings.batch')}
+                  </Text>
                   {!props.isPro && (
-                    <View style={styles.proLabel} paddingH-3 paddingV-1 br20 marginB-4 marginL-5>
-                      <Text white default>{i18n.t('other.pro')}</Text>
+                    <View
+                      style={styles.proLabel}
+                      paddingH-3
+                      paddingV-1
+                      br20
+                      marginB-4
+                      marginL-5
+                    >
+                      <Text white default>
+                        {i18n.t('other.pro')}
+                      </Text>
                     </View>
                   )}
                 </View>
                 <SettingSwitch
                   value={props.settings.batch}
-                  onChange={!props.isPro ? props.goPricingPage : props.setSettingValue}
+                  onChange={
+                    !props.isPro ? props.goPricingPage : props.setSettingValue
+                  }
                   name="batch"
                 />
               </View>
-              <Text gray marginR-50 defaultLight>{i18n.t('screens.settings.batchDescription')}</Text>
+              <Text gray marginR-50 defaultLight>
+                {i18n.t('screens.settings.batchDescription')}
+              </Text>
             </View>
             <View marginB-25>
               <View row spread centerV>
-                <Text h2 marginB-5 style={styles.textNormal}>{i18n.t('screens.settings.vibrate')}</Text>
+                <Text h2 marginB-5 style={styles.textNormal}>
+                  {i18n.t('screens.settings.vibrate')}
+                </Text>
                 <SettingSwitch
                   value={props.settings.vibrate}
                   onChange={props.setSettingValue}
                   name="vibrate"
                 />
               </View>
-              <Text gray marginR-50 defaultLight>{i18n.t('screens.settings.vibrateDescription')}</Text>
+              <Text gray marginR-50 defaultLight>
+                {i18n.t('screens.settings.vibrateDescription')}
+              </Text>
             </View>
             <View marginB-25>
               <View row spread centerV>
-                <Text h2 marginB-5 style={styles.textNormal}>{i18n.t('screens.settings.beep')}</Text>
+                <Text h2 marginB-5 style={styles.textNormal}>
+                  {i18n.t('screens.settings.beep')}
+                </Text>
                 <SettingSwitch
                   value={props.settings.beep}
                   onChange={props.setSettingValue}
                   name="beep"
                 />
               </View>
-              <Text gray marginR-50 defaultLight>{i18n.t('screens.settings.beepDescription')}</Text>
+              <Text gray marginR-50 defaultLight>
+                {i18n.t('screens.settings.beepDescription')}
+              </Text>
             </View>
             <View marginB-25>
               <View row spread centerV>
-                <Text h2 marginB-5 style={styles.textNormal}>{i18n.t('screens.settings.beep')}</Text>
+                <Text h2 marginB-5 style={styles.textNormal}>
+                  {i18n.t('screens.settings.beep')}
+                </Text>
                 <SettingSwitch
                   value={props.settings.history}
                   onChange={props.setSettingValue}
@@ -203,11 +285,15 @@ export default function SettingsView(props: Props) {
                   name="history"
                 />
               </View>
-              <Text gray marginR-50 defaultLight>{i18n.t('screens.settings.historyDescription')}</Text>
+              <Text gray marginR-50 defaultLight>
+                {i18n.t('screens.settings.historyDescription')}
+              </Text>
             </View>
             <View marginB-25>
               <View row spread centerV>
-                <Text h2 marginB-5 style={styles.textNormal}>{i18n.t('screens.settings.duplicate')}</Text>
+                <Text h2 marginB-5 style={styles.textNormal}>
+                  {i18n.t('screens.settings.duplicate')}
+                </Text>
                 <SettingSwitch
                   value={props.settings.duplicate}
                   onChange={props.setSettingValue}
@@ -215,14 +301,22 @@ export default function SettingsView(props: Props) {
                   name="duplicate"
                 />
               </View>
-              <Text gray marginR-50 defaultLight>{i18n.t('screens.settings.duplicateDescription')}</Text>
+              <Text gray marginR-50 defaultLight>
+                {i18n.t('screens.settings.duplicateDescription')}
+              </Text>
             </View>
           </View>
         </View>
         <View row marginB-25>
           <View flex marginR-20>
-            <TouchableOpacity onPress={() => Linking.openURL('mailto:apps@insider.io?subject=QRCode')}>
-              <Text center gray default>{i18n.t('screens.settings.contact')}</Text>
+            <TouchableOpacity
+              onPress={() =>
+                Linking.openURL('mailto:apps@insider.io?subject=QRCode')
+              }
+            >
+              <Text center gray default>
+                {i18n.t('screens.settings.contact')}
+              </Text>
             </TouchableOpacity>
           </View>
           <View flex>
@@ -231,12 +325,14 @@ export default function SettingsView(props: Props) {
                 center
                 gray
                 default
-                onPress={() => Rate.rate({
-                  AppleAppID: '1445350234',
-                  GooglePackageName: 'io.insider.apps.qr',
-                  preferInApp: true,
-                  openAppStoreIfInAppFails: true,
-                })}
+                onPress={() =>
+                  Rate.rate({
+                    AppleAppID: '1445350234',
+                    GooglePackageName: 'io.insider.apps.qr',
+                    preferInApp: true,
+                    openAppStoreIfInAppFails: true,
+                  })
+                }
               >
                 {i18n.t('screens.settings.rate')}
               </Text>
