@@ -10,19 +10,16 @@ const enhancers = [
     thunkMiddleware,
     createLogger({
       collapsed: true,
-      // eslint-disable-next-line no-undef
       predicate: () => __DEV__,
     }),
   ),
 ];
 
-/* eslint-disable no-undef */
-const composeEnhancers = (
-  __DEV__ &&
-  typeof (window) !== 'undefined' &&
-  window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
-) || compose;
-/* eslint-enable no-undef */
+const composeEnhancers =
+  (__DEV__ &&
+    typeof window !== 'undefined' &&
+    window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) ||
+  compose;
 
 const enhancer = composeEnhancers(...enhancers);
 
@@ -33,9 +30,5 @@ const persistConfig = {
 };
 
 const persistedReducer = persistReducer(persistConfig, reducer);
-export const store = createStore(
-  persistedReducer,
-  {},
-  enhancer,
-);
+export const store = createStore(persistedReducer, {}, enhancer);
 export const persistor = persistStore(store);
