@@ -1,8 +1,4 @@
-import {
-  StyleSheet,
-  StatusBar,
-  Platform,
-} from 'react-native';
+import { StyleSheet, StatusBar, Platform } from 'react-native';
 
 import colors from './colors';
 
